Cambios en ActorControllerTest

+++ Catalogo-Pruebas/src/main/java/com/example/application/controllers/ActorController.java
@@ -24,7 +24,6 @@ import com.example.exceptions.BadRequestException;
 import com.example.exceptions.DuplicateKeyException;
 import com.example.exceptions.InvalidDataException;
 import com.example.exceptions.NotFoundException;
-import com.fasterxml.jackson.core.JsonProcessingException;
 
 import jakarta.validation.Valid;
 
@@ -44,17 +43,21 @@ public class ActorController {
 	// http://localhost:8001/actores/get
 	@GetMapping(path = "/get")
 	public @ResponseBody List<ActorShort> getActors(@RequestParam(required = false) String sort)
-			throws JsonProcessingException {
+			throws Exception {
 		if (sort != null)
 			return (List<ActorShort>) actService.getByProjection(Sort.by(sort), ActorShort.class);
+		if(sort == null)
+		{
+			throw new Exception();
+		}
 		return actService.getByProjection(ActorShort.class);
 	}
 
 	// http://localhost:8001/actores/get/2
 	@GetMapping(path = "/get/{id}")
-	public ActorDTO getOneActor(@PathVariable int id) throws NotFoundException {
+	public ActorDTO getOneActor(@PathVariable int id) throws Exception {
 		var item = actService.getOne(id);
-		if (item.isEmpty()) {
+		if (item.isEmpty() || id <0) {
 			throw new NotFoundException();
 		}
 		return ActorDTO.from(item.get());
@@ -62,7 +65,11 @@ public class ActorController {
 
 	// http://localhost:8001/actores/peliculasDelActor/2
 	@GetMapping(path = "/peliculasDelActor/{id}")
-	public List<ElementoDto<Integer, String>> getPeliculasFromActor(@PathVariable int id) throws NotFoundException {
+	public List<ElementoDto<Integer, String>> getPeliculasFromActor(@PathVariable int id) throws Exception {
+		if(id < 0)
+		{
+			throw new Exception();
+		}
 		return actService.getOne(id).get().getFilmActors().stream()
 				.map(f -> new ElementoDto<>(f.getFilm().getFilmId(), f.getFilm().getTitle())).toList();
 	}
@@ -72,6 +79,10 @@ public class ActorController {
 	public @ResponseBody Actor addNewActor(@RequestParam String firstname, @RequestParam String lastname)
 			throws InvalidDataException, org.springframework.dao.DuplicateKeyException, DuplicateKeyException {
 
+		if(firstname == null | lastname == null)
+		{
+			throw new InvalidDataException();
+		}
 		var actorAux = new Actor();
 		actorAux.setActorId(0);
 		actorAux.setFirstName(firstname.toUpperCase());
@@ -87,7 +98,11 @@ public class ActorController {
 	// http://localhost:8001/actores/201
 	// {"id":201,"nombre":"KK","apellidos": "KKK"}
 	@DeleteMapping(path = "/{id}")
-	public void delete(@PathVariable int id) {
+	public void delete(@PathVariable int id) throws InvalidDataException {
+		if(id <0)
+		{
+			throw new InvalidDataException("ERROR. La id no puede ser menor a 0 (Controller)");
+		}
 		actService.deleteById(id);
 	}
